remove useless stuff in constants

+++ src/util/constants.rs
@@ -18,7 +18,7 @@ pub const INSTANCE_SHIFT: isize = SEQUENCE_LENGTH;
 pub const SEQUENCE_SHIFT: isize = 0;
 
 pub const DISTRIBUTED_SLEEP_TIME: Duration =
-    Duration::from_nanos((10u64.pow(9) / RESET_INCREMENT as u64) as u64);
+    Duration::from_nanos(10u64.pow(9) / RESET_INCREMENT as u64);
 
 // Snowflake constants
 pub const MAX_TIMESTAMP_MILLIS: i64 = (1 << 41) - 1;
